retail_data_analysis/requirements.txt
ZODB
pymongo
mysql-connector-python
pandas
numpy
matplotlib
scikit-learn
